// src/domain/enums/habito.enums.ts
export enum EnumHabitColor {
  RED = 'RED',
  BLUE = 'BLUE',
  GREEN = 'GREEN',
  YELLOW = 'YELLOW',
  PURPLE = 'PURPLE',
  PINK = 'PINK',
}

export enum EnumHabitIcon {
  WEIGHTS = 'WEIGHTS',
  BOOK = 'BOOK',
  SAVE = 'SAVE',
  MEDITATION = 'MEDITATION',
  RUNNING = 'RUNNING',
  BIKE = 'BIKE',
  HANGER = 'HANGER',
  CUTLERY = 'CUTLERY',
  FONT = 'FONT',
  WATER = 'WATER',
  MOON = 'MOON',
  CAR = 'CAR',
  PENCIL = 'PENCIL',
  CODE = 'CODE',
}